🐼 Merge and fix conflict

+++ src/main/java/com/ticket/server/dtos/CustomerDtos/AddCustomerRequest.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public class AddCustomerRequest {
     private String name;
     private String identifyNum;
-    private int birthday;
+    private Long birthday;
     private String phone;
     private String email;
     private Gender gender;
