+++ src/main/java/com/ticket/server/entities/CustomerEntity.java
@@ -18,7 +18,7 @@ public class CustomerEntity {
 
     private String name;
     private String identifyNum;
-    private int birthday;
+    private Long birthday;
     private String phone;
     private String email;
